SDE-1045: connected the provider to the autosde-client
Added manageiq-providers-autosde to the Gemfile, and pointed it to the local plugins folder.
Autosde manager now requires the autosde-client.
When verifying credentials for creating a new autosde-manager instance, it tries to login with the test ibmid user.
Moved the autosde-client to our namespace, Changed dict format to older syntax so idea works better.